Bugfixes in gossipmenu and alot of refactoring.
Bugfixes in gossipmenu and alot of refactoring.

Now shows a popup when clicking an enchant, which shows description and price.

Changed the enchant price label in menu.

Changed npc_text for each menu-section.

+++ src/PriceHelper.h
@@ -3,7 +3,7 @@
 #include "EnchantManager.h"
 
 // Define currency conversions
-#define COPPER(n) ((uint64)(n)) // Convert n to copper
+#define COPPER(n) ((uint32)(n)) // Convert n to copper
 #define SILVER(n) (COPPER(n) * 100) // Convert n to silver
 #define GOLD(n)   (SILVER(n) * 100) // Convert n to gold
